split files for managing the store
