+++ package.json
@@ -23,6 +23,7 @@
     "eslint-plugin-react": "7.20.0",
     "jest": "26.0.1",
     "react-test-renderer": "16.13.1",
+    "redux-mock-store": "1.5.4",
     "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.11.0"
