+++ src/components/artist-question-screen/artist-question-screen.e2e.test.js
@@ -2,6 +2,8 @@ import React from 'react';
 import Enzyme, {mount} from 'enzyme';
 import {MemoryRouter} from 'react-router-dom';
 import Adapter from 'enzyme-adapter-react-16';
+import configureStore from 'redux-mock-store';
+import {Provider} from 'react-redux';
 import {ArtistQuestionScreen} from './artist-question-screen';
 import questions from '../../mocks/test-questions';
 
@@ -13,14 +15,21 @@ describe(`ArtistQuestionScreen`, () => {
   describe(`onAnswerClick`, () => {
     test(`When use clicks answer with "Thad Jones" as an artist,
       onAcserClick should be called with "artist" and "Thad Jones" as parameters`, () => {
+      const mockStore = configureStore([]);
+      const store = mockStore({
+        mistakes: 0
+      });
       const onAnswerClick = jest.fn();
+
       const wrapper = mount(
-          <MemoryRouter>
-            <ArtistQuestionScreen
-              question={questions[0]}
-              onAnswerClick={onAnswerClick}
-            />
-          </MemoryRouter>
+          <Provider store={store}>
+            <MemoryRouter>
+              <ArtistQuestionScreen
+                question={questions[0]}
+                onAnswerClick={onAnswerClick}
+              />
+            </MemoryRouter>
+          </Provider>
       );
       wrapper
         .find(`input[value="Thad Jones"]`)
@@ -31,14 +40,21 @@ describe(`ArtistQuestionScreen`, () => {
 
     test(`When use clicks answer with "Miles Davis" as an artist,
       onAcserClick should be called with "artist" and "Miles Davis" as parameters`, () => {
+      const mockStore = configureStore([]);
+      const store = mockStore({
+        mistakes: 0
+      });
       const onAnswerClick = jest.fn();
+
       const wrapper = mount(
-          <MemoryRouter>
-            <ArtistQuestionScreen
-              question={questions[0]}
-              onAnswerClick={onAnswerClick}
-            />
-          </MemoryRouter>
+          <Provider store={store}>
+            <MemoryRouter>
+              <ArtistQuestionScreen
+                question={questions[0]}
+                onAnswerClick={onAnswerClick}
+              />
+            </MemoryRouter>
+          </Provider>
       );
       wrapper
         .find(`input[value="Miles Davis"]`)
