+++ src/components/game-mistakes/game-mistakes.js
@@ -6,7 +6,7 @@ const GameMistakes = (props) => {
   const wrongs = [];
   for (let i = 0; i < props.mistakes; i++) {
     wrongs.push(
-        <div className="wrong"></div>
+        <div className="wrong" key={i}></div>
     );
   }
   return (
