+++ src/components/wellcome-screen/wellcome-screen.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import {Link} from 'react-router-dom';
-import {incStepAction} from '../../reducer';
+import {incStepAction} from '../../action-creators/action-creators';
 
 const WellcomeScreen = (props) => {
   const {mistakesLimit, onStartClick} = props;
