+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {createStore} from 'redux';
 import {Provider} from 'react-redux';
-import {reducer, initialState} from './reducer';
+import {reducer, initialState} from './reducers/reducer';
 import App from './components/app/app';
 import questions from './mocks/questions';
 import settings from './mocks/settings';
